ai4bharat_stt_server/requirements.txt
fastapi
pydantic
torch
numpy
uvicorn
transformers
torchaudio 
onnx 
onnxruntime-gpu  # Note: onnxruntime-gpu replaces onnxruntime, don't install both
pyarr
loguru>=0.7.3